clear stations

+++ service/todo.go
@@ -3,6 +3,8 @@ package service
 import (
 	"context"
 	"database/sql"
+	"fmt"
+	"strings"
 
 	"github.com/TechBowl-japan/go-stations/model"
 )
@@ -26,7 +28,27 @@ func (s *TODOService) CreateTODO(ctx context.Context, subject, description strin
 		confirm = `SELECT subject, description, created_at, updated_at FROM todos WHERE id = ?`
 	)
 
-	return nil, nil
+	res, err := s.db.ExecContext(ctx, insert, subject, description)
+	if err != nil {
+		return nil, err
+	}
+
+	lastID, err := res.LastInsertId()
+	if err != nil {
+		return nil, err
+	}
+
+	todo := model.TODO{
+		ID: lastID,
+	}
+
+	if err := s.db.
+		QueryRowContext(ctx, confirm, lastID).
+		Scan(&todo.Subject, &todo.Description, &todo.CreatedAt, &todo.UpdatedAt, &todo.DeletedAt); err != nil {
+		return nil, err
+	}
+
+	return &todo, nil
 }
 
 // ReadTODO reads TODOs on DB.
@@ -36,7 +58,33 @@ func (s *TODOService) ReadTODO(ctx context.Context, prevID, size int64) ([]*mode
 		readWithID = `SELECT id, subject, description, created_at, updated_at FROM todos WHERE id < ? ORDER BY id DESC LIMIT ?`
 	)
 
-	return nil, nil
+	var rows *sql.Rows
+	var err error
+	if prevID == 0 {
+		if size == 0 {
+			size = 3
+		}
+		rows, err = s.db.QueryContext(ctx, read, size)
+	} else {
+		rows, err = s.db.QueryContext(ctx, readWithID, prevID, size)
+	}
+
+	if err == sql.ErrNoRows {
+		return nil, &model.ErrNotFound{
+			Message: fmt.Sprintf("received ID %d does not exist in the database", prevID),
+		}
+	}
+
+	defer rows.Close()
+
+	var todos []*model.TODO
+	for rows.Next() {
+		todo := model.TODO{}
+		rows.Scan(&todo.ID, &todo.Subject, &todo.Description, &todo.CreatedAt, &todo.UpdatedAt)
+		todos = append(todos, &todo)
+	}
+
+	return todos, nil
 }
 
 // UpdateTODO updates the TODO on DB.
@@ -46,12 +94,59 @@ func (s *TODOService) UpdateTODO(ctx context.Context, id int64, subject, descrip
 		confirm = `SELECT subject, description, created_at, updated_at FROM todos WHERE id = ?`
 	)
 
-	return nil, nil
+	res, err := s.db.ExecContext(ctx, update, subject, description, id)
+	if err != nil {
+		return nil, err
+	}
+
+	rows, err := res.RowsAffected()
+	if err != nil {
+		return nil, err
+	}
+
+	if rows == 0 {
+		return nil, &model.ErrNotFound{
+			Message: fmt.Sprintf("received ID %d does not exist in the database", id),
+		}
+	}
+	todo := model.TODO{
+		ID: id,
+	}
+	err = s.db.QueryRowContext(ctx, confirm, id).Scan(&todo.Subject, &todo.Description, &todo.CreatedAt, &todo.UpdatedAt)
+	if err != nil {
+		return nil, err
+	}
+	return &todo, nil
 }
 
 // DeleteTODO deletes TODOs on DB by ids.
 func (s *TODOService) DeleteTODO(ctx context.Context, ids []int64) error {
-	const deleteFmt = `DELETE FROM todos WHERE id IN (?%s)`
+	if len(ids) == 0 {
+		return nil
+	}
+	const deleteQueryTemplate = `DELETE FROM todos WHERE id IN (?%s)`
+	query := fmt.Sprintf(deleteQueryTemplate, strings.Repeat(", ?", len(ids)-1))
+	contextArgs := make([]interface{}, len(ids))
+
+	for i, id := range ids {
+		contextArgs[i] = id
+	}
+
+	res, err := s.db.ExecContext(ctx, query, contextArgs...)
+	if err != nil {
+		return err
+	}
+
+	rows, err := res.RowsAffected()
+	if err != nil {
+		return err
+	}
+
+	if rows == 0 {
+		return &model.ErrNotFound{
+			Message: fmt.Sprintf("received ID %d does not exist in the database", ids),
+		}
+	}
 
 	return nil
 }
